Move test to `test_types.py`
